+++ packages/oruga/src/components/dropdown/DropdownItem.vue
@@ -1,12 +1,11 @@
-<script setup lang="ts" generic="T">
-import { computed, onMounted, type PropType } from "vue";
+<script setup lang="ts" generic="T = string">
+import { computed, type ComputedRef, type PropType } from "vue";
 
 import { getOption } from "@/utils/config";
-import { uuid } from "@/utils/helpers";
-import { defineClasses } from "@/composables";
+import { uuid, isEqual } from "@/utils/helpers";
+import { defineClasses, useProviderChild } from "@/composables";
 
-import { injectDropdown } from "./useDropdownShare";
-
+import type { DropdownComponent } from "./types";
 import type { ComponentClass, DynamicComponent } from "@/types";
 
 /**
@@ -78,32 +77,26 @@ const emits = defineEmits<{
     (e: "click", value: T, event: Event): void;
 }>();
 
-// inject parent dropdown component if used inside one
-const { parentDropdown } = injectDropdown<T>();
-
-onMounted(() => {
-    if (!parentDropdown.value)
-        throw new Error("You should wrap oDropdownItem on a oDropdown");
-});
+// Inject functionalities and data from the parent component
+const { parent } = useProviderChild<ComputedRef<DropdownComponent<T>>>();
 
 const isClickable = computed(
-    () =>
-        !parentDropdown.value.props.disabled &&
-        !props.disabled &&
-        props.clickable,
+    () => !parent.value.props.disabled && !props.disabled && props.clickable,
 );
 
 const isActive = computed(() => {
-    if (parentDropdown.value.selected === null) return false;
-    if (parentDropdown.value.props.multiple)
-        return parentDropdown.value.selected.indexOf(props.value as T) >= 0;
-    return props.value === parentDropdown.value.selected;
+    if (parent.value.selected === null) return false;
+    if (parent.value.props.multiple && Array.isArray(parent.value.selected))
+        return parent.value.selected.some((selected) =>
+            isEqual(props.value, selected),
+        );
+    return isEqual(props.value, parent.value.selected);
 });
 
 /** Click listener, select the item. */
 function selectItem(event: Event): void {
     if (!isClickable.value) return;
-    parentDropdown.value.selectItem(props.value as T);
+    parent.value.selectItem(props.value as T);
     emits("click", props.value as T, event);
 }
 
@@ -115,7 +108,7 @@ const rootClasses = defineClasses(
         "itemDisabledClass",
         "o-drop__item--disabled",
         null,
-        computed(() => parentDropdown.value.props.disabled || props.disabled),
+        computed(() => parent.value.props.disabled || props.disabled),
     ],
     ["itemActiveClass", "o-drop__item--active", null, isActive],
     ["itemClickableClass", "o-drop__item--clickable", null, isClickable],